temp commit before creating motor_controller class

+++ sprokyt_firmware/src/motor/motor_controller.h
@@ -3,7 +3,9 @@
 #ifndef _MOTOR_CONTROLLER_H_
 #define _MOTOR_CONTROLLER_H_
 
+#include <stdint.h>
+
 void Init_MotorController(void);
-void SetMotor(int motorIndxMask, float power);
+void SetMotor(uint8_t motorIndxMask, float power);
 
 #endif // _MOTOR_CONTROLLER_H_
